Move ZonedDateTimeConverter to common
