+++ cab-app-common/src/main/java/by/arvisit/cabapp/common/util/CommonConstants.java
@@ -1,0 +1,11 @@
+package by.arvisit.cabapp.common.util;
+
+import java.time.ZoneId;
+
+public final class CommonConstants {
+
+    public static final ZoneId EUROPE_MINSK_TIMEZONE = ZoneId.of("Europe/Minsk");
+
+    private CommonConstants() {
+    }
+}
