+++ cab-app-rides-service/src/main/java/by/arvisit/cabapp/ridesservice/util/AppConstants.java
@@ -1,12 +1,9 @@
 package by.arvisit.cabapp.ridesservice.util;
 
-import java.time.ZoneId;
-
 public final class AppConstants {
 
     public static final int MAX_SCORE = 5;
-    public static final ZoneId EUROPE_MINSK_TIMEZONE = ZoneId.of("Europe/Minsk");
-    
+
     private AppConstants() {
     }
 }
